Move page finding logic into container
